+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Added
+
+- Add optional `output.localeCodesModule` config file setting which generates a
+  TypeScript module that exports `sourceLocale`, `targetLocales`, and
+  `allLocales` using the locale codes from your config file. Use for keeping
+  your config file and client config in sync.
 
 ## [0.3.0] - 2020-08-25